Parser initial

+++ src/Benchmarks/Image/Svg/SvgGradientStop.hpp
@@ -14,7 +14,7 @@ namespace Elpida
 
 	private:
 		SvgColor _color;
-		float _offset;
+		double _offset;
 	};
 
 } // Elpida
